requirements.txt
yt-dlp
openai-whisper
ollama
torch
setuptools-rust
python-dotenv
pydantic
ffmpeg-python
fastapi
uvicorn
